requirements.txt
numpy
streamlit

pillow
